requirements.txt
Flask==1.1.2
Flask_Cors==3.0.10
Pillow==8.4.0
psycopg2==2.9.2
PyJWT==2.3.0
pytest==3.10.1
requests==2.21.0
Werkzeug==1.0.1
